+++ src/main/java/roomescape/infrastructure/payment/paymentResponseErrorHandler.java
@@ -1,4 +1,4 @@
-package roomescape.config;
+package roomescape.infrastructure.payment;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
